enhance

+++ LlamaIndex/test/clean_chroma_collection.py
@@ -1,5 +1,8 @@
 import sys, chromadb
 
 db = chromadb.PersistentClient(path="LlamaIndex/chroma_db")
-collection = len(sys.argv) == 2 and sys.argv[1] or "local"
-db.delete_collection(collection)
+collection = len(sys.argv) == 2 and sys.argv[1] or None
+if collection is None:
+    print("provide the collectino name")
+else:
+    db.delete_collection(collection)
